toolkit/requirements.txt
flask==3.1.*
gunicorn==23.*
requests==2.32.*
pyyaml==6.*
